refactor: make user inferface use presenters
Until now, the UI classes have been rather messy, in including a lot of
business logic within them.
This commit refactors the classes into view class and presents, mostly
because it becomes clear that that is the way to go with the new editor,
since that one uses interchangeable models.